Abbreviate days of week and months to three letters

+++ ui/src/lib.ts
@@ -14,9 +14,9 @@ const formatDate: Function = (date: Date): string => {
 
 const dateFormat = (selectedEventDate: Date) =>
   `${selectedEventDate.toLocaleString("default", {
-    weekday: "long",
+    weekday: "short",
   })}, ${selectedEventDate.toLocaleString("default", {
-    month: "long",
+    month: "short",
   })} ${selectedEventDate.getDate()}`;
 
 const timeFormat = (selectedEventDate: Date) =>
